react client, concurrently, admin key, delete product from cart, etc

+++ src/routes/product.js
@@ -1,6 +1,8 @@
 const express = require('express');
 const router = express.Router();
 const Product = require('../models/product')
+const User = require('../models/user')
+const auth = require('../middleware/auth')
 
 router.get('/products', async (req, res) => {
     try {
@@ -25,7 +27,7 @@ router.get('/products/:id', async (req, res) => {
     try {
         const product = await Product.findById(req.params.id);
         if(!product) {
-            res.status(404).send({error: "Product not found"})
+            return res.status(404).send({error: "Product not found"})
         }
         res.send(product)
     } catch (e) {
@@ -33,14 +35,21 @@ router.get('/products/:id', async (req, res) => {
     }
 })
 
-//hay que permitir borrarlo si es admin. por ahora elimina siempre
-router.delete('/products/:id', async (req, res) => {
+//elimina producto solo si es admin
+router.delete('/products/:id', auth, async (req, res) => {
     try {
-        const product = await Product.findByIdAndDelete(req.params.id);
-        if(!product) {
-            res.status(404).send({error: "Product not found"})
+        const user = await User.findById(req.user.id);
+        console.log(user)
+        if(user.admin) {
+            const product = await Product.findByIdAndDelete(req.params.id);
+            if(!product) {
+            return res.status(404).send({error: "Product not found"})
+            }
+            res.send(product)
+        } else {
+            res.status(403).send({error: Unauthorized})
         }
-        res.send(product)
+        
     } catch (e) {
         res.status(404).send(e)
     }
